fix bug in node.Run and reorg osc-based gens
